requirements.txt
streamlit
langchain
langchain_community
huggingface-hub
faiss-cpu
ctransformers
google-generativeai
googletrans==4.0.0-rc1
pathlib
pypdf
httpcore==0.15.0
